Added static_cast #4478

+++ src/udp_engine.cpp
@@ -368,7 +368,7 @@ void zmq::udp_engine_t::sockaddr_to_msg (zmq::msg_t *msg_,
 
     char port[6];
     // gets the number of digits in the sin_port
-    const size_t sin_port_len = snprintf (0, 0, "%d", ntohs(addr_->sin_port));
+    const size_t sin_port_len = snprintf (0, 0, "%d", static_cast<int> (ntohs(addr_->sin_port)));
     const int port_len =
       snprintf (port, sin_port_len, "%d", static_cast<int> (ntohs(addr_->sin_port)));
     zmq_assert (port_len > 0);
